backend/api/llm.py
from fastapi import APIRouter, HTTPException
from pydantic import BaseModel
import logging
# TODO: Import LLM service functions and potentially database functions for context

logger = logging.getLogger(__name__)
router = APIRouter()

class LLMPrompt(BaseModel):
    book_id: str
    text: str # Selected text or context
    prompt: str # User's question/request
    # TODO: Add fields for position/section reference

# TODO: Implement endpoints for LLM interactions:
# POST /llm/summarize
# POST /llm/ask
# etc.

# Example placeholder:
# @router.post("/ask")
# async def ask_llm(llm_prompt: LLMPrompt):
#     logger.info(f"Received LLM question for book {llm_prompt.book_id}")
#     # TODO: Fetch context from book data in DB
#     # TODO: Call LLM service
#     # response = await llm_service.ask_question(llm_prompt.prompt, llm_prompt.text, context)
#     response = "Placeholder LLM response."
#     # TODO: Optionally save interaction as a note
#     return {"response": response}
